refactored more
